[Rename] region 파일 이동 #214
